Added concept of maps and arrays
Before there was no difference between the two, they were all collections, however and Empty array and an empty object are two different things, so this commit makes sure they are represented in the dapt packet as two different things. I was considering just returning "null" on an empty object or array, but again, that is a different thing.