Refactored how btokens work
While working on things borrow checking rules kept rendering the btoken
useless. So instead of saving a reference of a slice, we just keep the
index of the token within the binary slice. This allows us to use
lifetimes so the compiler understands where the underlying data is a
reference from.